UI to for the cards in position
I did some of the functionalities for the position system yesterday accidentally, instead of the UI so now I am adding what I needed to. Note another member did the "Create Position" card thinking it was there part in the review.html which is why I did not do it.